Properly handle add and delete at the end
If a chunk deletes or adds to the end it's possible for the index to
reference past the end of the chuck.  Since these are positions to
add/delete it makes no sense to try to print them.  Skip printing of the
content at locations indicated to avoid index out of bounds error.

This also revealed that when there's an uneven amount of chunks the
resulting None chuncks were not handled properly.  Check for None chunks
and treat them as add/delete at the end.

Fixes #2.